mid-project-347359563/services/seat-service/requirements.txt
fastapi
uvicorn
aio-pika
pika
pymysql
sqlalchemy
python-dotenv
pydantic
requests
cryptography